CC-010-RENAME (#60)
* refactor: rename .hpp to .h

* deleted old files
