+++ console/main.cpp
@@ -1,4 +1,4 @@
-#include "app.hpp"
+#include "app.h"
 
 int main() {
     return 0;
