+++ console/menu_action.cpp
@@ -1,4 +1,4 @@
-#include "menu_action.hpp"
+#include "menu_action.h"
 
 namespace menu {
 
@@ -10,15 +10,11 @@ struct Snapshot;
 /** @todo Command argument parsing functions */
 
 void PrintSnapshot(std::ostream& output, Snapshot&& snapshot) {}
-    
-} // namespace detail
 
-CommandHandlers::CommandHandlers(Menu& menu,
-                                 std::ostream& output,
-                                 client_lib::IOrderBookClient& client)
-    : menu_(menu),
-      output_(output),
-      client_(client) {}
+}  // namespace detail
+
+CommandHandlers::CommandHandlers(Menu& menu, std::ostream& output, client_lib::IOrderBookClient& client) :
+    menu_(menu), output_(output), client_(client) {}
 
 bool CommandHandlers::Buy(Menu::CommandArgs&& args) const {
     return false;
@@ -40,8 +36,7 @@ bool CommandHandlers::Help() const {
     return false;
 }
 
-ArgsParsers::ArgsParsers(Menu& menu, std::istream& input)
-    : menu_(menu), input_(input) {}
+ArgsParsers::ArgsParsers(Menu& menu, std::istream& input) : menu_(menu), input_(input) {}
 
 Menu::CommandArgs ArgsParsers::ParseBuyArgs(std::istream&) {
     return {};
@@ -55,4 +50,4 @@ Menu::CommandArgs ArgsParsers::ParseBookArgs(std::istream&) {
     return {};
 }
 
-} // namespace menu
+}  // namespace menu
